requirements.txt
aiogram==3.*
fastapi==0.*
uvicorn[standard]==0.*
python-dotenv==1.*
aiohttp==3.*
